autenticación de monitores

+++ ProyectoNaturAdventureWeb/src/es/uji/ei1027/naturAdventure/controller/InstructorController.java
@@ -1,5 +1,7 @@
 package es.uji.ei1027.naturAdventure.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import es.uji.ei1027.naturAdventure.dao.InstructorDao;
 import es.uji.ei1027.naturAdventure.domain.Instructor;
+import es.uji.ei1027.naturAdventure.domain.UserDetails;
 
 @Controller
 @RequestMapping("/instructor")
@@ -24,19 +27,35 @@ public class InstructorController {
 	}
 	
 	@RequestMapping("/list")
-	public String listInstructors( Model model ) {
-		model.addAttribute( "instructors", instructorDao.getInstructors() );
-		return "instructor/list";
+	public String listInstructors( Model model, HttpSession session ) {
+		if( checkAuthentified( session ) ) {			
+			model.addAttribute( "instructors", instructorDao.getInstructors() );
+			return "instructor/list";
+		}
+		model.addAttribute( "user", new UserDetails() );
+		session.setAttribute( "nextURL", "/instructor/list.html" );
+		return "login";
+		
 	}
 	
 	@RequestMapping(value="/add")
-	public String addInstructor( Model model ) {
-		model.addAttribute( "instructor", new Instructor() );
-		return "instructor/add";
+	public String addInstructor( Model model, HttpSession session ) {
+		if( checkAuthentified( session ) ) {
+			model.addAttribute( "instructor", new Instructor() );
+			return "instructor/add";	
+		}
+		model.addAttribute( "user", new UserDetails() );
+		session.setAttribute( "nextURL", "/instructor/add.html" );
+		return "login";
 	}
 	
 	@RequestMapping(value="/add", method=RequestMethod.POST)
-	public String processAddSubmit( @ModelAttribute("instructor") Instructor instructor, BindingResult bindingResult ) {
+	public String processAddSubmit( @ModelAttribute("instructor") Instructor instructor, BindingResult bindingResult, HttpSession session, Model model ) {
+		if( !checkAuthentified( session ) ) {
+			model.addAttribute( "user", new UserDetails() );
+			session.setAttribute( "nextURL", "/instructor/add.html" );
+			return "login";
+		}
 		if( bindingResult.hasErrors() ) {
 			return "instructor/add";
 		}
@@ -45,13 +64,23 @@ public class InstructorController {
 	}
 	
 	@RequestMapping(value="/update/{nif}")
-	public String updateInstructor( Model model, @PathVariable String nif ) {
-		model.addAttribute( "instructor" , instructorDao.getInstructor( nif ) );
-		return "instructor/update";
+	public String updateInstructor( Model model, @PathVariable String nif, HttpSession session ) {
+		if( checkAuthentified( session ) ) {
+			model.addAttribute( "instructor" , instructorDao.getInstructor( nif ) );
+			return "instructor/update";
+		}
+		model.addAttribute( "user", new UserDetails() );
+		session.setAttribute( "nextURL", "/instructor/update/" + nif + ".html" );
+		return "login";
 	}
 	
 	@RequestMapping(value="/update/{nif}", method=RequestMethod.POST)
-	public String processUpdateSubmit( @ModelAttribute("instructor") Instructor instructor, BindingResult bindingResult ) {
+	public String processUpdateSubmit( @ModelAttribute("instructor") Instructor instructor, BindingResult bindingResult, HttpSession session, Model model ) {
+		if( !checkAuthentified( session ) ) {
+			model.addAttribute( "user", new UserDetails() );
+			session.setAttribute( "nextURL", "/instructor/update/" + instructor.getNif() + ".html" );
+			return "login";
+		}
 		if( bindingResult.hasErrors() ) {
 			return "instructor/update";
 		}
@@ -60,11 +89,23 @@ public class InstructorController {
 	}
 	
 	@RequestMapping(value="/delete/{nif}")
-	public String deleteInstructor( @PathVariable String nif ) {
+	public String deleteInstructor( @PathVariable String nif, HttpSession session, Model model ) {
+		if( !checkAuthentified( session ) ) {
+			model.addAttribute( "user", new UserDetails() );
+			session.setAttribute( "nextURL", "/instructor/delete/" + nif + ".html" );
+			return "login";
+		}
 		Instructor instructor = new Instructor();
 		instructor.setNif( nif );
 		instructorDao.deleteInstructor( instructor );
 		return "redirect:../list.html";
+	}
+	
+	private boolean checkAuthentified( HttpSession session ) {
+		if( session.getAttribute( "user" ) == null ) {
+			return false;
+		}
+		return true;
 	}
 
 }
